Small updates to executable file documentation

+++ pyrcareworld/pyrcareworld/envs/base_env.py
@@ -15,7 +15,7 @@ class RCareWorld(ABC):
     rcareworld base environment class.
 
     Args:
-        executable_file: Str, the absolute path of Unity executable file. None for last used executable file; "@editor" for using Unity Editor.
+        executable_file: Str, the absolute path of Unity executable file. None uses config.json; "@editor" for using Unity Editor.
         scene_file: Str, the absolute path of Unity scene JSON file. All JSON files locate at `StraemingAssets/SceneData` by default.
         assets: List, the list of pre-load assets. All assets in the list will be pre-loaded in Unity when the environment is initialized, which will save time during instanciating.
         graphics: Bool, True for showing GUI and False for headless mode.
@@ -70,7 +70,7 @@ class RCareWorld(ABC):
             PROC_TYPE = "release"
             self.port = self.port + 1 + proc_id  # default release port
         else:  # error
-            raise ValueError(f"Executable file {executable_file} not exists")
+            raise ValueError(f"Executable file {executable_file} does not exist")
 
         self.communicator = RFUniverseCommunicator(
             port=self.port,
@@ -232,9 +232,9 @@ class RCareWorld(ABC):
         """
         Close the environment
         """
-        if self.process is not None:
+        if hasattr(self, "process") and self.process is not None:
             self.process.kill()
-        if self.communicator is not None:
+        if hasattr(self, "communicator") and self.communicator is not None:
             self.communicator.close()
 
     # def SetAutoSimulate(self, auto: bool):
